refactor: separate Python binding models and encoding
