+++ native-lib/python/pytest.ini
@@ -1,6 +1,6 @@
 [pytest]
 testpaths = tests
-addopts = -m "not tck"
+addopts = --import-mode=importlib -m "not tck"
 markers =
     unit: fast tests that do not require the native DataWeave library
     integration: tests that execute the staged native DataWeave library
